# Enable Parallel test execution in pkg/scheduler