--- a/pkg/scheduler/schedule_one_test.go
+++ b/pkg/scheduler/schedule_one_test.go
@@ -319,6 +319,7 @@ func (t *TestPlugin) Filter(ctx context.Context, state *framework.CycleState, po
 }
 
 func TestSchedulerMultipleProfilesScheduling(t *testing.T) {
+	t.Parallel()
 	nodes := []runtime.Object{
 		st.MakeNode().Name("node1").UID("node1").Obj(),
 		st.MakeNode().Name("node2").UID("node2").Obj(),
@@ -445,6 +446,7 @@ func TestSchedulerMultipleProfilesScheduling(t *testing.T) {
 }
 
 func TestSchedulerScheduleOne(t *testing.T) {
+	t.Parallel()
 	testNode := v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "node1", UID: types.UID("node1")}}
 	client := clientsetfake.NewSimpleClientset(&testNode)
 	eventBroadcaster := events.NewBroadcaster(&events.EventSinkImpl{Interface: client.EventsV1()})
@@ -639,6 +641,7 @@ func TestSchedulerScheduleOne(t *testing.T) {
 }
 
 func TestSchedulerNoPhantomPodAfterExpire(t *testing.T) {
+	t.Parallel()
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	queuedPodStore := clientcache.NewFIFO(clientcache.MetaNamespaceKeyFunc)
@@ -704,6 +707,7 @@ func TestSchedulerNoPhantomPodAfterExpire(t *testing.T) {
 }
 
 func TestSchedulerNoPhantomPodAfterDelete(t *testing.T) {
+	t.Parallel()
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	queuedPodStore := clientcache.NewFIFO(clientcache.MetaNamespaceKeyFunc)
@@ -773,6 +777,7 @@ func TestSchedulerNoPhantomPodAfterDelete(t *testing.T) {
 }
 
 func TestSchedulerFailedSchedulingReasons(t *testing.T) {
+	t.Parallel()
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	queuedPodStore := clientcache.NewFIFO(clientcache.MetaNamespaceKeyFunc)
@@ -855,6 +860,7 @@ func TestSchedulerFailedSchedulingReasons(t *testing.T) {
 }
 
 func TestSchedulerWithVolumeBinding(t *testing.T) {
+	t.Parallel()
 	findErr := fmt.Errorf("find err")
 	assumeErr := fmt.Errorf("assume err")
 	bindErr := fmt.Errorf("bind err")
@@ -1001,6 +1007,7 @@ func TestSchedulerWithVolumeBinding(t *testing.T) {
 }
 
 func TestSchedulerBinding(t *testing.T) {
+	t.Parallel()
 	table := []struct {
 		podName      string
 		extenders    []framework.Extender
@@ -1036,7 +1043,9 @@ func TestSchedulerBinding(t *testing.T) {
 	}
 
 	for _, test := range table {
+		test := test
 		t.Run(test.name, func(t *testing.T) {
+			t.Parallel()
 			pod := st.MakePod().Name(test.podName).Obj()
 			defaultBound := false
 			client := clientsetfake.NewSimpleClientset(pod)
@@ -1084,6 +1093,7 @@ func TestSchedulerBinding(t *testing.T) {
 }
 
 func TestUpdatePod(t *testing.T) {
+	t.Parallel()
 	tests := []struct {
 		name                     string
 		currentPodConditions     []v1.PodCondition
@@ -1225,7 +1235,9 @@ func TestUpdatePod(t *testing.T) {
 		},
 	}
 	for _, test := range tests {
+		test := test
 		t.Run(test.name, func(t *testing.T) {
+			t.Parallel()
 			actualPatchRequests := 0
 			var actualPatchData string
 			cs := &clientsetfake.Clientset{}
@@ -1263,6 +1275,7 @@ func TestUpdatePod(t *testing.T) {
 }
 
 func TestSelectHost(t *testing.T) {
+	t.Parallel()
 	tests := []struct {
 		name          string
 		list          framework.NodeScoreList
@@ -1310,7 +1323,9 @@ func TestSelectHost(t *testing.T) {
 	}
 
 	for _, test := range tests {
+		test := test
 		t.Run(test.name, func(t *testing.T) {
+			t.Parallel()
 			// increase the randomness
 			for i := 0; i < 10; i++ {
 				got, err := selectHost(test.list)
@@ -1332,6 +1347,7 @@ func TestSelectHost(t *testing.T) {
 }
 
 func TestFindNodesThatPassExtenders(t *testing.T) {
+	t.Parallel()
 	tests := []struct {
 		name                  string
 		extenders             []st.FakeExtender
@@ -1483,7 +1499,9 @@ func TestFindNodesThatPassExtenders(t *testing.T) {
 	}
 
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
 			var extenders []framework.Extender
 			for ii := range tt.extenders {
 				extenders = append(extenders, &tt.extenders[ii])
@@ -1511,6 +1529,7 @@ func TestFindNodesThatPassExtenders(t *testing.T) {
 }
 
 func TestSchedulerSchedulePod(t *testing.T) {
+	t.Parallel()
 	fts := feature.Features{}
 	tests := []struct {
 		name               string
@@ -1976,7 +1995,9 @@ func TestSchedulerSchedulePod(t *testing.T) {
 		},
 	}
 	for _, test := range tests {
+		test := test
 		t.Run(test.name, func(t *testing.T) {
+			t.Parallel()
 			cache := internalcache.New(time.Duration(0), wait.NeverStop)
 			for _, pod := range test.pods {
 				cache.AddPod(pod)
@@ -2125,10 +2146,12 @@ func TestFindFitSomeError(t *testing.T) {
 	}
 
 	for _, node := range nodes {
+		node := node
 		if node.Name == pod.Name {
 			continue
 		}
 		t.Run(node.Name, func(t *testing.T) {
+			t.Parallel()
 			status, found := diagnosis.NodeToStatusMap[node.Name]
 			if !found {
 				t.Errorf("failed to find node %v in %v", node.Name, diagnosis.NodeToStatusMap)
@@ -2209,6 +2232,7 @@ func TestFindFitPredicateCallCounts(t *testing.T) {
 //     is the one being scheduled.
 //   - don't get the same score no matter what we schedule.
 func TestZeroRequest(t *testing.T) {
+	t.Parallel()
 	// A pod with no resources. We expect spreading to count it as having the default resources.
 	noResources := v1.PodSpec{
 		Containers: []v1.Container{
@@ -2297,7 +2321,9 @@ func TestZeroRequest(t *testing.T) {
 	}
 
 	for _, test := range tests {
+		test := test
 		t.Run(test.name, func(t *testing.T) {
+			t.Parallel()
 			client := clientsetfake.NewSimpleClientset()
 			informerFactory := informers.NewSharedInformerFactory(client, 0)
 
@@ -2357,6 +2383,7 @@ func TestZeroRequest(t *testing.T) {
 var lowPriority, midPriority, highPriority = int32(0), int32(100), int32(1000)
 
 func TestNumFeasibleNodesToFind(t *testing.T) {
+	t.Parallel()
 	tests := []struct {
 		name                     string
 		percentageOfNodesToScore int32
@@ -2398,7 +2425,9 @@ func TestNumFeasibleNodesToFind(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
 			sched := &Scheduler{
 				percentageOfNodesToScore: tt.percentageOfNodesToScore,
 			}
@@ -2453,6 +2482,7 @@ func TestFairEvaluationForNodes(t *testing.T) {
 }
 
 func TestPreferNominatedNodeFilterCallCounts(t *testing.T) {
+	t.Parallel()
 	tests := []struct {
 		name                  string
 		pod                   *v1.Pod
@@ -2479,7 +2509,9 @@ func TestPreferNominatedNodeFilterCallCounts(t *testing.T) {
 	}
 
 	for _, test := range tests {
+		test := test
 		t.Run(test.name, func(t *testing.T) {
+			t.Parallel()
 			// create three nodes in the cluster.
 			nodes := makeNodeList([]string{"node1", "node2", "node3"})
 			client := clientsetfake.NewSimpleClientset(test.pod)