<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?

/kind feature

<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind feature

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->

#### What this PR does / why we need it:

Since go v1.7, the go testing package provides the ability to run [tests in parallel](https://pkg.go.dev/testing#T.Parallel). As this [post](https://engineering.mercari.com/en/blog/entry/20220408-how_to_use_t_parallel/) describes, unit tests will run in parallel at the package level. Therefore, this change allows increasing the number of processes during the execution of SIG/scheduling unit tests

* Before (forcing testing cleanup):
```bash
$ for i in {1..3}; do make test WHAT=./pkg/scheduler/; done
+++ [0914 12:42:23] Building go targets for linux/amd64
    k8s.io/kubernetes/hack/make-rules/helpers/go2make (non-static)
+++ [0914 12:42:25] Running tests without code coverage and with -race
ok      k8s.io/kubernetes/pkg/scheduler 6.445s
+++ [0914 12:42:34] Building go targets for linux/amd64
    k8s.io/kubernetes/hack/make-rules/helpers/go2make (non-static)
+++ [0914 12:42:37] Running tests without code coverage and with -race
ok      k8s.io/kubernetes/pkg/scheduler 6.424s
+++ [0914 12:42:46] Building go targets for linux/amd64
    k8s.io/kubernetes/hack/make-rules/helpers/go2make (non-static)
+++ [0914 12:42:48] Running tests without code coverage and with -race
ok      k8s.io/kubernetes/pkg/scheduler 6.461s
```
* After (forcing testing cleanup):
```bash
$ for i in {1..3}; do make test WHAT=./pkg/scheduler/; done
+++ [0914 12:43:59] Building go targets for linux/amd64
    k8s.io/kubernetes/hack/make-rules/helpers/go2make (non-static)
+++ [0914 12:44:01] Running tests without code coverage and with -race
ok      k8s.io/kubernetes/pkg/scheduler 1.160s
+++ [0914 12:44:05] Building go targets for linux/amd64
    k8s.io/kubernetes/hack/make-rules/helpers/go2make (non-static)
+++ [0914 12:44:08] Running tests without code coverage and with -race
ok      k8s.io/kubernetes/pkg/scheduler 1.160s
+++ [0914 12:44:12] Building go targets for linux/amd64
    k8s.io/kubernetes/hack/make-rules/helpers/go2make (non-static)
+++ [0914 12:44:14] Running tests without code coverage and with -race
ok      k8s.io/kubernetes/pkg/scheduler 1.159s
```



#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
NA

#### Special notes for your reviewer:

I have to force the testing cache to make sure improvements are achieved
```bash
$ git diff hack/make-rules/test.sh
diff --git a/hack/make-rules/test.sh b/hack/make-rules/test.sh
index d58bc929752..01465b9be7b 100755
--- a/hack/make-rules/test.sh
+++ b/hack/make-rules/test.sh
@@ -257,6 +257,7 @@ runTests() {
 
   verifyPathsToPackagesUnderTest "$@"
 
+  go clean -testcache
   # If we're not collecting coverage, run all requested tests with one 'go test'
   # command, which is much faster.
   if [[ ! ${KUBE_COVER} =~ ^[yY]$ ]]; then
```

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note

```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```