@electrocucaracha 2022-09-08
> Just a question, what's the difference with test option "test.parallel"? I am fine with adding this to speed up the test execution, but this will force to check all those testcases instead of just eventhandlers.

[AFAIK](https://stackoverflow.com/a/45974915/2727227), the `-parallel` and `-test.parallel` flags controls the maximum number of tests to run simultaneously. The `t.Parallel()` method signals that test to be run in parallel with another parallel tests.