@aojea 2022-09-08
I'm not sure we should do this, there are some tests that depend on global variables, the ones that use metrics per example ... you can check a list of test that can not run in parallel here, this job runs the same test 10 times, and some of them fail because depend on some global state https://testgrid.k8s.io/sig-testing-canaries#ci-kubernetes-generate-make-test-count10

I'm -1 on implementing this until:
- we have a clear indication of the improvement, i.e. reduce CI execution in X%, 
- we have those tests green on https://testgrid.k8s.io/sig-testing-canaries#ci-kubernetes-generate-make-test-count10