@electrocucaracha 2022-09-08
@aojea I agree with you, this needs to be a better analysis or take a look case by case. Unfortunately, in order to see improvements, a majority of test cases needs to implement this parallelism (kind of chicken and egg situation). Anyway, we can consider this approach for future testing implementations, and leave the existing ones for further analysis.