@alculquicondor 2022-09-08
The tests touched in this PR seem fine to run in parallel.

Do you have proof of improvements in time execution?