@electrocucaracha 2022-09-08
> The tests touched in this PR seem fine to run in parallel.
> 
> Do you have proof of improvements in time execution?

Maybe I need to do better analysis because my results are giving me similar performance numbers.

```shell
$ make clean
+++ [0908 12:34:02] Verifying Prerequisites....             
+++ [0908 12:34:03] Removing _output directory
$ GOMAXPROCS=2 time make test WHAT=./pkg/scheduler/
+++ [0908 12:34:07] Building go targets for linux/amd64     
    k8s.io/kubernetes/hack/make-rules/helpers/go2make (non-static)
+++ [0908 12:34:17] Building go targets for linux/amd64  
    k8s.io/code-generator/cmd/prerelease-lifecycle-gen (non-static)
+++ [0908 12:34:21] Generating prerelease lifecycle code for 27 targets
+++ [0908 12:34:23] Building go targets for linux/amd64
    k8s.io/code-generator/cmd/deepcopy-gen (non-static)
+++ [0908 12:34:26] Generating deepcopy code for 242 targets
+++ [0908 12:34:31] Building go targets for linux/amd64      
    k8s.io/code-generator/cmd/defaulter-gen (non-static)
+++ [0908 12:34:32] Generating defaulter code for 95 targets   
+++ [0908 12:34:39] Building go targets for linux/amd64               
    k8s.io/code-generator/cmd/conversion-gen (non-static)
+++ [0908 12:34:42] Generating conversion code for 132 targets                                                                                               
+++ [0908 12:34:56] Building go targets for linux/amd64           
    k8s.io/kube-openapi/cmd/openapi-gen (non-static)      
+++ [0908 12:35:08] Generating openapi code for KUBE
+++ [0908 12:35:27] Generating openapi code for AGGREGATOR
+++ [0908 12:35:28] Generating openapi code for APIEXTENSIONS
+++ [0908 12:35:29] Generating openapi code for CODEGEN
+++ [0908 12:35:30] Generating openapi code for SAMPLEAPISERVER
+++ [0908 12:35:31] Running tests without code coverage and with -race
ok      k8s.io/kubernetes/pkg/scheduler 6.467s
277.09user 27.85system 2:46.29elapsed 183%CPU (0avgtext+0avgdata 939796maxresident)k
0inputs+3395216outputs (24148major+14145651minor)pagefaults 0swaps
$ git checkout add_tparallel-scheduling
Switched to branch 'add_tparallel-scheduling'
$ make clean           
+++ [0908 12:39:27] Verifying Prerequisites....
+++ [0908 12:39:27] Removing _output directory
$ GOMAXPROCS=2 time make test WHAT=./pkg/scheduler/
+++ [0908 12:39:35] Building go targets for linux/amd64
    k8s.io/kubernetes/hack/make-rules/helpers/go2make (non-static)
+++ [0908 12:39:44] Building go targets for linux/amd64
    k8s.io/code-generator/cmd/prerelease-lifecycle-gen (non-static)
+++ [0908 12:39:49] Generating prerelease lifecycle code for 27 targets
+++ [0908 12:39:51] Building go targets for linux/amd64
    k8s.io/code-generator/cmd/deepcopy-gen (non-static)
+++ [0908 12:39:53] Generating deepcopy code for 242 targets
+++ [0908 12:39:58] Building go targets for linux/amd64
    k8s.io/code-generator/cmd/defaulter-gen (non-static)
+++ [0908 12:40:00] Generating defaulter code for 95 targets
+++ [0908 12:40:07] Building go targets for linux/amd64
    k8s.io/code-generator/cmd/conversion-gen (non-static)
+++ [0908 12:40:09] Generating conversion code for 132 targets
+++ [0908 12:40:23] Building go targets for linux/amd64
    k8s.io/kube-openapi/cmd/openapi-gen (non-static)
+++ [0908 12:40:35] Generating openapi code for KUBE
+++ [0908 12:40:53] Generating openapi code for AGGREGATOR
+++ [0908 12:40:54] Generating openapi code for APIEXTENSIONS
+++ [0908 12:40:56] Generating openapi code for CODEGEN
+++ [0908 12:40:57] Generating openapi code for SAMPLEAPISERVER
+++ [0908 12:40:58] Running tests without code coverage and with -race
ok      k8s.io/kubernetes/pkg/scheduler 6.273s
281.89user 27.54system 2:47.10elapsed 185%CPU (0avgtext+0avgdata 955840maxresident)k
0inputs+3395216outputs (24155major+14169833minor)pagefaults 0swaps
```