@chendave 2022-09-09
> Maybe I need to do better analysis because the results that I'm getting are similar.

since you just update the eventhandler, so if you want to check the improvement, you should focus on eventhandler's testcases, I tried to run this with `schedule_one_test.go`, it does improve one second.