@chendave 2022-09-09
On master:

go test k8s.io/kubernetes/pkg/scheduler
ok      k8s.io/kubernetes/pkg/scheduler 6.115s


With this change, 
go test k8s.io/kubernetes/pkg/scheduler
ok      k8s.io/kubernetes/pkg/scheduler 5.914s

note that I removed the test of `extender_test.go` and `scheduler_test.go`.