@electrocucaracha 2022-09-09
Makes sense @chendave , I'm going to try one by one. BTW, I'm trying to reduce the issues reported by [this linter](https://github.com/kunwardeep/paralleltest)