@electrocucaracha 2022-09-12
Ok, I'll break down this PR in changes per file to manage it better