@electrocucaracha 2022-09-14
@aojea @alculquicondor @chendave According to my metrics, these changes reduces ~5.5x the testing time in the tests declared under pkg/scheduler folder. I couldn't apply parallelism in the following sections:

* pkg/scheduler/extender_test.go:274
* pkg/scheduler/schedule_one_test.go:545
* pkg/scheduler/schedule_one_test.go:954
* pkg/scheduler/schedule_one_test.go:2074
* pkg/scheduler/schedule_one_test.go:2113
* pkg/scheduler/schedule_one_test.go:2167
* pkg/scheduler/schedule_one_test.go:2185
* pkg/scheduler/schedule_one_test.go:2441