@alculquicondor 2022-09-14
can you share some numbers?