@electrocucaracha 2022-09-14
> can you share some numbers?

@alculquicondor I included them in the PR description, but basically they went from ~6.443 sec to ~1.16 sec.