@aojea 2022-09-14
> these changes reduces ~5.5x the testing time in the tests declared under pkg/scheduler folder.

now we are talking :smile: :muscle: