@aojea 2022-09-14
the verify job failure is legit, you have some gofmt errors